web_demo/requirements.txt
torchvision
torch
numpy
Pillow
flask_restful
flask
matplotlib
